Added back light and shadows
Looks a bit off though. Light and shadow on the sphere should be
inverted. And I still have artifacts.

Oh, and I noticed that the position of the sphere and triangle looks
fine when you view everything from far away. As soon as you move between
the sphere and triangle, things bugs out.