Temporarily removed the color stage
I think this will make things easier to debug. Right now, the issue is
that the entire mesh is not being rendered. I think the issue comes from
that the data might be layed in memory the wrong way. Or that I've
somehow misunderstood how big the ssbo needs to be.